+++ otserv/branches/revbattlesys/weapons.cpp
@@ -447,7 +447,7 @@ int32_t Weapon::getManaCost(const Player* player) const
 
 	if(manaPercent != 0){
 		int32_t currentMana = player->getMana();
-		int32_t manaCost = currentMana * (((double)manaPercent) / 100);
+		int32_t manaCost = (currentMana * manaPercent)/100;
 		return manaCost;
 	}
